fixed refresh method and refactored outh
