+++ server.js
@@ -21,6 +21,7 @@ const {
   NODE_ENV,
 } = require('./dotenvValidator');
 const { registerAdmin } = require('./tests/helper-functions');
+const RuleError = require('./errors/rule.error');
 
 connectDB();
 
@@ -47,20 +48,25 @@ const server = new ApolloServer({
       message: `method: ${req.method}/baseUrl: ${req.baseUrl}/date:${req.fresh}/`,
     });
     if (token) {
-      const { userId } = verifyUser(token);
+      try {
+        const { userId } = verifyUser(token);
 
-      if (!userId) {
-        logger.error({
-          level: 'error',
-          message: formatErrorForLogger(INVALID_PERMISSIONS),
-        });
-        return null;
+        if (!userId) {
+          logger.error({
+            level: 'error',
+            message: formatErrorForLogger(INVALID_PERMISSIONS),
+          });
+          return null;
+        }
+        return {
+          user: await userService.getUserByFieldOrThrow('_id', userId),
+        };
+      } catch (e) {
+        return new RuleError(e.message, e.statusCode);
       }
-      return {
-        user: await userService.getUserByFieldOrThrow('_id', userId),
-      };
     }
   },
+
   plugins: [errorOutputPlugin],
   formatError,
   introspection: true,
